football-betting-platform/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
pymysql>=1.1.0
cryptography>=41.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
coverage>=7.0.0
pytest-cov>=4.1.0
